+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
  - @eps macro
  - polyintersect()
-
+ - rand(bb::BoundingBox)
+ - in(pt, bb) alias isinside(pt, bb)
+ 
 ### Changed
 
 ### Removed